HARVESTER: Use product model to override common model
